feat: throw exception when task manager cannot retrieve/remove task
Add throwing exception to get methods of an in-memory task manager
in order to isolate data processing from error handling and avoid
returning null to the caller. Also add throwing exception to remove
methods of the manager to let the caller know that the entity does
not exist.

Refactor add methods and update methods of the manager to make the
code more concise.