fix: calculate last used id on the fly for file backed task manager
Move a file backed task manager from storing the last used id in a
file to calculating it from task ids 'on the fly' while loading the
file. It makes the code clearer while the chance that a user keeps
ids of removed tasks (for any reason) is low.

Refactor tests for snapshots saved by a file backed task manager,
and add the explicit file header to them like it is in tests for
snapshots loaded.