feat: sort tasks by start time, check that tasks do not overlap
Add a method to retrieve tasks and subtasks sorted by their start
time, i.e. prioritized from most urgent to less urgent (for tasks
and subtasks with start time only). Add a check that tasks and
subtasks do not overlap. Merge a task duration and start time
consistency check with the above overlapping tasks check.

Move a check that id of a new task, a new epic, a new subtask
does not already exist in a task manager from a file backed task
manager to an in-memory task manager.