fix: use Duration in getter/setter; check duration at service layer
Move task duration getter and setter to using java.time.Duration
instead of Long (=number of minutes) in order to make it easier to
work with the value (no conversion needed). Add truncating of task
duration down to minutes (at a service layer). Move a check that
duration is at least 1 minute to a service layer. Move truncating
of task start time down to minutes to a service layer, too.

Add tests for values loaded from a file:
- Id is of an integer type,
- Duration (in minutes) is of an integer type,
- Epic id for a subtask is of an integer type.