fix: allow set null as id/epic id for model classes, refactor tests
Change setters for model classes to accept null as id (for of a task,
an epic, a subtask) and epic id (for a subtask) - all business logic
should be moved to a service layer.

Add tests that setters of model classes accept null.

Fix typos in AbstractTaskManagerTest and TestModels classes.