fix: rename task manager methods after CRUD operations
Rename task manager methods in order them to match basic CRUD
operations: 'create' instead of 'add' and 'delete' instead of
'remove'.

Add 'ById' suffix to the names of methods that retrieve a single
task/epic/subtask: getTaskById, getEpicById, getSubtaskById. It
is a good practice because model classes may have a number of
identifiers.