fix: create task with custom id via update, not create
Move creating of a task (an epic, a subtask) with a custom id
from a 'create' method to an 'update' method as this operation
is idempotent (i.e. it has the same effect when applied several
times as when applied only once).

Re-use code of an 'update' method in a 'create' method leaving
just the following in the latter:
- Assigning the next id to the task being created.
- Returning that id to the caller.

Refactor code for updating the list of prioritized tasks.