feat: return caller created/updated model when create/update
Return a created (updated) task, epic, subtask from methods that
create or update a task, an epic, a subtask. It will be in line
with HTTP specification where a server should return the caller
the representation of a created (modified) resource.

Add new utility methods for test subtask creating.

Move checks that a custom ID updates the last used ID to separate
tests.

Update project properties in order to limit the scope where JUnit
methods are available to project testing.